Move ChangeHash to protocol crate
