+++ automerge-backend/src/lib.rs
@@ -30,12 +30,12 @@ mod serialize;
 mod time;
 mod actor_map;
 
-pub use automerge_protocol::{ActorID};
+pub use automerge_protocol::{ActorID, ChangeHash};
 
 pub use crate::patch::{DiffEdit, Patch, Diff, MapDiff, SeqDiff, ObjDiff};
 pub use crate::protocol::{
     Change, ChangeRequest, ChangeRequestType, DataType, Key, ObjType, ObjectID,
-    OpType, Operation, Value, ChangeHash
+    OpType, Operation, Value
 };
 pub use backend::Backend;
 pub use error::AutomergeError;
