+++ automerge-backend/src/serialize.rs
@@ -1,7 +1,7 @@
 use crate::patch::{Diff, DiffEdit, MapDiff, ObjDiff, SeqDiff};
 use crate::protocol::{
     DataType, ElementID, Key, ObjType, ObjectID, OpID, OpType, Operation, ReqOpType, RequestKey,
-    UndoOperation, Value, ChangeHash
+    UndoOperation, Value
 };
 use serde::de;
 use serde::de::{Error, MapAccess, Unexpected, Visitor};
@@ -10,26 +10,6 @@ use serde::{Deserialize, Deserializer, Serialize, Serializer};
 use std::collections::HashMap;
 use std::fmt;
 use std::str::FromStr;
-
-impl Serialize for ChangeHash {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        hex::encode(&self.0).serialize(serializer)
-    }
-}
-
-impl<'de> Deserialize<'de> for ChangeHash {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let s = String::deserialize(deserializer)?;
-        let vec = hex::decode(&s).map_err(|_| de::Error::invalid_value(de::Unexpected::Str(&s), &"A valid hex string"))?;
-        Ok(vec.as_slice().into())
-    }
-}
 
 impl Serialize for Diff {
     fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
