+++ automerge-protocol/src/lib.rs
@@ -1,3 +1,7 @@
 mod actor_id;
+mod change_hash;
+
+mod serde;
 
 pub use actor_id::ActorID;
+pub use change_hash::ChangeHash;
