+++ automerge-protocol/src/serde/mod.rs
@@ -1,0 +1,1 @@
+mod change_hash;
